move content of container to __init__.py
