+++ docs/source/examples/raw_sql.md
@@ -25,7 +25,7 @@ import sqlalchemy as sa
 
 from pydiverse.pipedag import Flow, Stage, materialize
 from pydiverse.pipedag.context import ConfigContext, StageLockContext
-from pydiverse.pipedag.container.container import RawSql
+from pydiverse.pipedag.container import RawSql
 from tests.fixtures.instances import with_instances
 
 """
