+++ src/pydiverse/pipedag/backend/lock/database.py
@@ -14,7 +14,7 @@ from pydiverse.pipedag.backend.lock.base import BaseLockManager, Lockable, LockS
 from pydiverse.pipedag.backend.table.sql.ddl import (
     CreateSchema,
 )
-from pydiverse.pipedag.container.container import Schema
+from pydiverse.pipedag.container import Schema
 from pydiverse.pipedag.errors import LockError
 
 DISABLE_DIALECT_REGISTRATION = "__DISABLE_DIALECT_REGISTRATION"
