+++ src/pydiverse/pipedag/backend/table/base.py
@@ -12,7 +12,7 @@ from typing_extensions import Self
 
 from pydiverse.pipedag import ConfigContext
 from pydiverse.pipedag._typing import T, TableHookResolverT
-from pydiverse.pipedag.container.container import RawSql, Table
+from pydiverse.pipedag.container import RawSql, Table
 from pydiverse.pipedag.context import RunContext, TaskContext
 from pydiverse.pipedag.errors import CacheError
 from pydiverse.pipedag.materialize.cache import TaskCacheInfo, lazy_table_cache_key
