+++ src/pydiverse/pipedag/backend/table/sql/dialects/postgres.py
@@ -18,8 +18,7 @@ from pydiverse.pipedag.backend.table.sql.hooks import (
 )
 from pydiverse.pipedag.backend.table.sql.sql import SQLTableStore
 from pydiverse.pipedag.backend.table.util import DType
-from pydiverse.pipedag.container import Table
-from pydiverse.pipedag.container.container import Schema
+from pydiverse.pipedag.container import Schema, Table
 from pydiverse.pipedag.materialize.details import (
     BaseMaterializationDetails,
     resolve_materialization_details_label,
