+++ src/pydiverse/pipedag/materialize/core.py
@@ -12,7 +12,7 @@ from typing import TYPE_CHECKING, Any, overload
 import sqlalchemy as sa
 
 from pydiverse.pipedag._typing import CallableT
-from pydiverse.pipedag.container.container import Blob, RawSql, Table
+from pydiverse.pipedag.container import Blob, RawSql, Table
 from pydiverse.pipedag.context import ConfigContext, RunContext, TaskContext
 from pydiverse.pipedag.context.context import (
     CacheValidationMode,
