+++ src/pydiverse/pipedag/materialize/debug.py
@@ -7,7 +7,7 @@ import structlog
 from pydiverse.pipedag import ConfigContext, Table
 from pydiverse.pipedag.backend import SQLTableStore
 from pydiverse.pipedag.backend.table.sql.ddl import DropTable
-from pydiverse.pipedag.container.container import Schema
+from pydiverse.pipedag.container import Schema
 from pydiverse.pipedag.context import TaskContext
 from pydiverse.pipedag.materialize.core import MaterializingTask
 from pydiverse.pipedag.materialize.store import mangle_table_name
