+++ src/pydiverse/pipedag/util/json.py
@@ -14,7 +14,7 @@ from functools import cache
 from pathlib import Path
 
 from pydiverse.pipedag import ConfigContext, Stage
-from pydiverse.pipedag.container.container import Blob, RawSql, Table
+from pydiverse.pipedag.container import Blob, RawSql, Table
 from pydiverse.pipedag.context import RunContext
 from pydiverse.pipedag.core.config import PipedagConfig
