+++ tests/test_cache/test_basic_cache_invalidation.py
@@ -5,7 +5,7 @@ import pytest
 import sqlalchemy as sa
 
 from pydiverse.pipedag import Blob, ConfigContext, Flow, Stage, Table
-from pydiverse.pipedag.container.container import RawSql
+from pydiverse.pipedag.container import RawSql
 from pydiverse.pipedag.context import StageLockContext
 from pydiverse.pipedag.context.context import CacheValidationMode
 from pydiverse.pipedag.materialize.core import (
