+++ tests/test_cache/test_ignore_cache_function.py
@@ -4,7 +4,7 @@ import pandas as pd
 import sqlalchemy as sa
 
 from pydiverse.pipedag import Blob, Flow, Stage, Table
-from pydiverse.pipedag.container.container import RawSql
+from pydiverse.pipedag.container import RawSql
 from pydiverse.pipedag.context import StageLockContext
 from pydiverse.pipedag.context.context import CacheValidationMode
 from pydiverse.pipedag.materialize.core import materialize
