+++ tests/test_flows/test_raw_sql_pipeline.py
@@ -5,7 +5,7 @@ from pathlib import Path
 import sqlalchemy as sa
 
 from pydiverse.pipedag import Flow, Stage, materialize
-from pydiverse.pipedag.container.container import RawSql
+from pydiverse.pipedag.container import RawSql
 from pydiverse.pipedag.context import ConfigContext, StageLockContext
 from tests.fixtures.instances import with_instances
